requirements.txt
fastapi==0.61.2
motor==2.2.0
pydantic[email]
python-decouple==3.3
uvicorn==0.11.8
python-jose==3.3.0
numpy
starlette
boto3
botocore
logger
sqlalchemy
uuid
jose
bcrypt
SQLAlchemy
PyMySQL
sqlalchemy_utils
cryptography
async-exit-stack 
async-generator
python-multipart
fastapi_utils
redis
pyyaml
pymongo==3.11.0
python-dotenv
pandas
kafka-python==2.0.2
aiohttp
requests
elasticsearch
vnstock==0.2.3
plotly
beautifulsoup4
pyjwt
passlib
authlib
httpx
itsdangerous
google-auth
mysql-connector-python
sendgrid